+++ packages/extension-chakra-store-locator/src/components/list.tsx
@@ -7,9 +7,9 @@
 
 import React, {useEffect, useState} from 'react'
 import {Accordion, AccordionItem, Box, Button} from '@chakra-ui/react'
-import {StoreLocatorListItem} from 'overridable!./store-locator-list-item'
+import {StoreLocatorListItem} from 'overridable!./list-item'
 import {useStoreLocator} from './use-store-locator'
-import {Stores, Store} from '../../types/store'
+import {Stores, Store} from '../types/store'
 
 interface StoreLocatorListProps {
     storesInfo?: Stores