+++ packages/extension-chakra-store-locator/src/pages/store-locator/index.tsx
@@ -7,7 +7,7 @@
 
 import React from 'react'
 import {Box, Container} from '@chakra-ui/react'
-import {StoreLocatorContent} from '../../components/store-locator/store-locator-content'
+import {StoreLocatorContent} from '../../components/content'
 
 const StoreLocatorPage = () => {
     return (