+++ packages/internal-lib-build/bin/internal-lib-build.js
@@ -31,9 +31,9 @@ const main = () => {
         sh.rm('-rf', './dist')
         sh.mkdir('./dist')
         execSync(
-            `${babel} --config-file ${babelConfig} src -x ".js",".jsx",".ts",".tsx" --ignore "**/test_fixtures/*","*.test.js","test.js" --out-dir dist --copy-files`
+            `"${babel}" --config-file "${babelConfig}" src -x ".js",".jsx",".ts",".tsx" --ignore "**/test_fixtures/*","*.test.js","test.js" --out-dir dist --copy-files`
         )
-        execSync(`node ${prepareDist}`)
+        execSync(`node "${prepareDist}"`)
     })
 
     program