+++ packages/pwa-kit-extension-sdk/src/shared/utils/resolver.test.js
@@ -14,14 +14,25 @@ describe('resolverUtils', () => {
             {
                 name: 'Correct paths are returned when overridable import is used in an application extension',
                 resourcePath: path.join('pages', 'sample'),
+                importSourceExtension: 'extension-a',
+                canonicalSource: path.join(
+                    process.cwd(),
+                    'node_modules',
+                    'extension-a',
+                    'src',
+                    'pages',
+                    'sample.tsx'
+                ),
                 extensions: [['extension-a'], ['extension-b'], ['extension-c']],
                 expected: [
+                    path.join(process.cwd(), 'app', 'overrides', 'extension-a', 'pages', 'sample'),
                     path.join(
                         process.cwd(),
                         'node_modules',
                         'extension-c',
                         'src',
                         'overrides',
+                        'extension-a',
                         'pages',
                         'sample'
                     ),
@@ -31,6 +42,7 @@ describe('resolverUtils', () => {
                         'extension-b',
                         'src',
                         'overrides',
+                        'extension-a',
                         'pages',
                         'sample'
                     ),
@@ -39,23 +51,33 @@ describe('resolverUtils', () => {
                         'node_modules',
                         'extension-a',
                         'src',
-                        'overrides',
                         'pages',
-                        'sample'
+                        'sample.tsx'
                     )
                 ]
             },
             {
                 name: 'Correct paths are returned when "some" extensions are disabled',
                 resourcePath: path.join('pages', 'sample'),
+                importSourceExtension: 'extension-a',
+                canonicalSource: path.join(
+                    process.cwd(),
+                    'node_modules',
+                    'extension-a',
+                    'src',
+                    'pages',
+                    'sample.tsx'
+                ),
                 extensions: [['extension-a'], ['extension-b'], ['extension-c', {enabled: false}]],
                 expected: [
+                    path.join(process.cwd(), 'app', 'overrides', 'extension-a', 'pages', 'sample'),
                     path.join(
                         process.cwd(),
                         'node_modules',
                         'extension-b',
                         'src',
                         'overrides',
+                        'extension-a',
                         'pages',
                         'sample'
                     ),
@@ -64,27 +86,51 @@ describe('resolverUtils', () => {
                         'node_modules',
                         'extension-a',
                         'src',
-                        'overrides',
                         'pages',
-                        'sample'
+                        'sample.tsx'
                     )
                 ]
             },
             {
                 name: 'Correct paths are returned when "all" extensions are disabled',
                 resourcePath: path.join('pages', 'sample'),
+                importSourceExtension: 'extension-a',
+                canonicalSource: path.join(
+                    process.cwd(),
+                    'node_modules',
+                    'extension-a',
+                    'src',
+                    'pages',
+                    'sample.tsx'
+                ),
                 extensions: [
                     ['extension-a', {enabled: false}],
                     ['extension-b', {enabled: false}],
                     ['extension-c', {enabled: false}]
                 ],
-                expected: []
+                expected: [
+                    path.join(process.cwd(), 'app', 'overrides', 'extension-a', 'pages', 'sample'),
+                    path.join(
+                        process.cwd(),
+                        'node_modules',
+                        'extension-a',
+                        'src',
+                        'pages',
+                        'sample.tsx'
+                    )
+                ]
             }
         ].forEach((testCase) => {
             test(`${testCase.name}`, () => {
-                const result = resolverUtils.buildCandidatePaths(testCase.resourcePath, {
-                    extensionEntries: testCase.extensions
-                })
+                const result = resolverUtils.buildCandidatePaths(
+                    testCase.resourcePath,
+                    testCase.importSourceExtension,
+                    {
+                        canonicalSource: testCase.canonicalSource,
+                        packageName: testCase.packageName,
+                        extensionEntries: testCase.extensions
+                    }
+                )
 
                 expect(result).toEqual(testCase.expected)
             })