+++ packages/commerce-sdk-react/CHANGELOG.md
@@ -1,4 +1,4 @@
-## v4.0.0-dev (Jun 21, 2024)
+## v3.1.1-preview.0 (Dec 02, 2024)
 
 ## v3.0.0 (Aug 07, 2024)
 - Add `meta.displayName` to queries. It can be used to identify queries in performance metrics or logs. [#1895](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1895)